Revert "Move Iperf3Monitor listening from Application class to MainActivityExtensions"
This reverts commit d26ba1a88ed6142f16e5f454f82df7406130707b.

things just doesn't seem to work good while it's not in the Application class :/
